+++ service/controller/resource/alert/services.go
@@ -1,4 +1,4 @@
-package servicemonitor
+package alert
 
 import (
 	"fmt"
